Response completed, minor edits in the article
All reviewer comments are addressed now and I formatted the table into
something more pretty and printed a PDF version of it.
In the article I made cosmetic changes for formatting, created a few
typos, and amended the abstract to reflect the 3 points of the main
message to make it stand out more.